Implement code changes to enhance functionality and improve performance
